long

+++ passport-main-manager/src/main/java/com/sogou/upd/passport/manager/account/impl/CommonManagerImpl.java
@@ -226,7 +226,7 @@ public class CommonManagerImpl implements CommonManager {
     @Override
     public boolean isCodeRight(String firstStr, int clientId, long ct, String originalCode) {
         String code = getCode(firstStr.toString(), clientId, ct);
-        int currentTime = (int)(System.currentTimeMillis()/1000);
+        long currentTime = System.currentTimeMillis()/1000;
         boolean isCodeEqual = code.equalsIgnoreCase(originalCode);
         boolean timeRight = ct > currentTime - CommonConstant.COOKIE_REQUEST_VAILD_TERM;
         if (isCodeEqual && timeRight) {
