+++ passport-main-web/src/main/java/com/sogou/upd/passport/web/account/api/SSOCookieController.java
@@ -85,7 +85,7 @@ public class SSOCookieController extends BaseController {
     private int getMaxAge(long et) {
         int maxAge = -1;
         if (et > 0) {
-            int currentTime = (int) (System.currentTimeMillis() / 1000);
+            long currentTime = System.currentTimeMillis() / 1000;
             maxAge = DateUtil.getIntervalSec(et, currentTime);
             if (maxAge == 0) {
                 maxAge = -1;
